gateway/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Google Cloud
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-tasks==2.15.0

# HTTP client
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Dev
pytest==7.4.4
pytest-asyncio==0.23.3
